refactor(metrics): move rollout_metrics.py back into metrics.py
Reclaim the conventional module name vacated by the old MetricsBuilder. Rename the source +
test (rollout_metrics.py -> metrics.py, test_rollout_metrics.py -> test_metrics.py) and fix
the stale compute_rollout_metrics doc references left from the train/eval split.
